SUPPORT.MD proposal
The CSS (Microsoft Customer Service and Support) team has a process
for providing support to some products or projects. To help improve
the experience for such cases, this update proposes a new SUPPORT.MD
file to be placed in the default repo templates with information for
maintainers to update and publish.

For questions in the proposal, please reach out to JHornick or comment
on this original pull request.